renaming lib.js to swarmAnimation, changing some dir structure
