+++ src/js/swarmAnimation.js
