--- test/js/DummySpec.js
@@ -1,10 +1,0 @@
-import swarmAnimate from '../../src/js/lib';
-
-
-describe("a first test boilerplate for swarmAnimate functions", function () {
-
-	it("has a dummy add one function", function () {
-		expect(swarmAnimate.addOneTest(1)).toEqual(2);
-	});
-
-});
