+++ test/specs/DummySpec.js
@@ -1,0 +1,10 @@
+import swarmAnimation from '../../src/js/swarmAnimation';
+
+
+describe('a first test boilerplate for swarmAnimate functions', function () {
+
+	it('has a dummy add one function', function () {
+		expect(swarmAnimation.addOneTest(1)).toEqual(2);
+	});
+
+});
